fix: address qodo findings on the shim ensure PR
- Share FakeLoginShellProbe via Capacitor.Tests.Helpers instead of
  duplicating it in the Core and App test suites (the repo rule: cross-suite
  helpers live in Helpers with a public surface).
- Off-macOS refusal (unsupported_platform) now beats an unknown probe in the
  classifier — the flow expects a stable platform row, not a probe-dependent
  one — and the daemon usage line lists the reviewer subcommand it dispatches.
